+++ frontend/src/App.jsx
@@ -1,0 +1,7 @@
+import Layout from "./layout/layout";
+
+function App() {
+  return <Layout />;
+}
+
+export default App;
