+++ frontend/src/components/header/Header.jsx
@@ -1,0 +1,7 @@
+import Navbar from "./navbar/Navbar";
+
+const Header = () => {
+  return <Navbar />;
+};
+
+export default Header;
