+++ frontend/src/components/header/navbar/registerButton.jsx
