+++ frontend/src/layout/layout.jsx
@@ -1,0 +1,13 @@
+import Header from "../components/header/Header";
+import Routes from "../routes/Routes";
+
+const Layout = () => {
+  return (
+    <>
+      <Header />
+      <Routes />
+    </>
+  );
+};
+
+export default Layout;
