+++ frontend/src/pages/auth/register/register.css
